+++ docs/CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.0] - 2023-29-11
+
+### Deprecated
+
+- !joined() command
+
+### Added
+
+- !whois() function to get detailed member information
+- Help message(s) for new function(s)
+
+### Fixed
+
+- Minor formatting
+
 ## [2.2.2] - 2023-29-11
 
 ### Removed